Extract & refactor mysqli TypeMapping (#262)
